ui: show version string in title menu
What a pain this was!
UI text objects (or layout objects, IDRK, though it seems more like the former) stop rendering
after they reach a certain width, and I haven't been able to get around this.

So I ended up cloning the copyright text (which is long enough), which means it's not a general
solution to the problem.